Update tests to rely on stable order of results
The 'compareResults' method that runs when using a snapshot as a source
was using proposalID lookup to compare and would show failures with
snapshots that tested proposals with malformed metadata such as missing
proposal IDs.

Updated this method to rely on the stable order to compare zipped
arrays of the actual and expected results.

Update the testWarningsAndErrors() method to use the extractEvolutionMetadata()
method instead of manually iterating through each proposal itself.

Make the readAndExtractProposalMetadata() method private since the tests
no longer need to perform extractions proposal by proposal.

Resolves #14